added logic and refactor

+++ src/main/java/com/myhotel/guestservice/repository/GuestDetailsRepository.java
@@ -3,5 +3,5 @@ package com.myhotel.guestservice.repository;
 import com.myhotel.guestservice.model.entity.GuestEntity;
 import org.springframework.data.repository.CrudRepository;
 
-public interface GuestDetailsRepository extends CrudRepository<GuestEntity, Integer> {
+public interface GuestDetailsRepository extends CrudRepository<GuestEntity, Long> {
 }
